nps_project/requirements.txt
psycopg2-binary==2.9.6
django-cors-headers==3.14.0
djangorestframework==3.14.0,<3.15.0
Faker>=17.0.0
djangorestframework-simplejwt>=5.2.2,<5.3.0
dj-rest_auth>=3.0.0,<3.1.0
python-dotenv==0.19.2
pytest==7.4.0
django-safedelete==1.3.1
PyYAML==6.0.1
ua-parser==0.18.0
user-agents==2.2.0
requests==2.31.0
coverage==7.2.5
